+++ NEWS.md
@@ -5,6 +5,8 @@
 ## Fixed
 * Fixed a bug in `ReadAndSplitVCFs` for merging adjacent SBSs into DBS when `variant.caller` is `mutect`.
 
+* Fixed a bug in`CheckAndRemoveDiscardedVariants` for removing wrong DBS variants.
+
 ## Updated
 * Updated function `CheckAndRemoveDiscardedVariants` to remove wrong DBS variants that have same base in the same position in REF and ALT (e.g. TA > TT or GT > CT).