+++ NEWS.md
@@ -1,4 +1,4 @@
-# ICAMS 2.3.1.90xx
+# ICAMS 2.3.1
 ## Added
 * Added an optional argument `files` in function `VCFsToZipFile`.